CLIM-1367: T17 Fix seasonal-vs-monthly rendering inconsistency + ReUse generatePeriodRangeLabel
